+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * (x/async) Rename futures into tasks, and future handlers into plugins
 * (x/async) Add a simple Plugin registry for tracking Plugins' IDs onchain
+* (x/async) Initial version of the QuantKit plugin
 
 ### Consensus Breaking Changes